+++ src/main/scala/io/viash/config/dependencies/Dependency.scala
@@ -25,7 +25,7 @@ import io.viash.ViashNamespace
 import io.viash.exceptions.MissingBuildYamlException
 
 @description(
-  """Specifies a Viash component (script or executable) that should be made available for the code defined in the functionality.
+  """Specifies a Viash component (script or executable) that should be made available for the code defined in the component.
     |The dependency components are collected and copied to the output folder during the Viash build step.
     |""".stripMargin)
 @exampleWithDescription(