+++ src/main/scala/io/viash/config/resources/JavaScriptScript.scala
@@ -26,8 +26,8 @@ import io.viash.config.Config
 import io.viash.config.arguments.{Argument, StringArgument, IntegerArgument, BooleanArgumentBase, LongArgument, DoubleArgument, FileArgument}
 
 @description("""An executable JavaScript script.
-               |When defined in functionality.resources, only the first entry will be executed when running the built component or when running `viash run`.
-               |When defined in functionality.test_resources, all entries will be executed during `viash test`.""".stripMargin)
+               |When defined in resources, only the first entry will be executed when running the built component or when running `viash run`.
+               |When defined in test_resources, all entries will be executed during `viash test`.""".stripMargin)
 @subclass("javascript_script")
 case class JavaScriptScript(
   path: Option[String] = None,