+++ src/main/scala/io/viash/config/resources/RScript.scala
@@ -26,8 +26,8 @@ import java.net.URI
 import io.viash.config.Config
 
 @description("""An executable R script.
-               |When defined in functionality.resources, only the first entry will be executed when running the built component or when running `viash run`.
-               |When defined in functionality.test_resources, all entries will be executed during `viash test`.""".stripMargin)
+               |When defined in resources, only the first entry will be executed when running the built component or when running `viash run`.
+               |When defined in test_resources, all entries will be executed during `viash test`.""".stripMargin)
 @subclass("r_script")
 case class RScript(
   path: Option[String] = None,