+++ src/main/scala/io/viash/config/resources/Resource.scala
@@ -26,7 +26,7 @@ import java.nio.file.NoSuchFileException
 import io.viash.schemas._
 
 @description(
-  """Resources are files that support the component. The first resource should be @[a script](scripting_languages) that will be executed when the functionality is run. Additional resources will be copied to the same directory.
+  """Resources are files that support the component. The first resource should be @[a script](scripting_languages) that will be executed when the component is run. Additional resources will be copied to the same directory.
     |
     |Common properties:
     |