+++ CHANGELOG.md
@@ -7,6 +7,8 @@
     config_mods: |
       .functionality.argument_groups[true].arguments[.multiple == true].multiple_sep := ":"
 
+* `functionality`: Remove the `functionality` layer from the config and move all fields to the top layer (PR #649).
+
 ## MINOR CHANGES
 
 * `package config`: Renamed `project config` to `package config` (PR #636). Now that we start using the config more, we came to the conclusion that "package" was better suited that "project".